// src/app/interfaces/clases.ts
export interface ClaseResponse {
    
    ok: boolean;
    _id: string;
    studentName: string;
    professorName: string;
    idMateria: string;
    nota1: string;
    nota2: string;
    nota3: string;

};

export interface Clase {
    
    _id: string;
    studentName: string;
    professorName: string;
    idMateria: string;
    nota1: string;
    nota2: string;
    nota3: string;
    nombreMateria: string;

};